+++ src/graphene_elastic/filter_backends/highlight/common.py
@@ -4,6 +4,7 @@ import graphene
 from graphene_elastic.types.json_string import JSONString
 
 from ...constants import DYNAMIC_CLASS_NAME_PREFIX
+from ...helpers import to_pascal_case
 from ..base import BaseBackend
 
 __title__ = 'graphene_elastic.filter_backends.ordering.common'
@@ -80,7 +81,7 @@ class HighlightFilterBackend(BaseBackend):
                         enum.Enum(
                             "{}{}{}BackendEnum".format(
                                 DYNAMIC_CLASS_NAME_PREFIX,
-                                self.prefix.title(),
+                                to_pascal_case(self.prefix),
                                 self.connection_field.type.__name__
                             ),
                             params