+++ src/graphene_elastic/filter_backends/source/common.py
@@ -4,6 +4,7 @@ import graphene
 
 from ..base import BaseBackend
 from ...constants import DYNAMIC_CLASS_NAME_PREFIX
+from ...helpers import to_pascal_case
 
 __title__ = 'graphene_elastic.filter_backends.ordering.common'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
@@ -54,7 +55,7 @@ class SourceFilterBackend(BaseBackend):
                         enum.Enum(
                             "{}{}{}BackendEnum".format(
                                 DYNAMIC_CLASS_NAME_PREFIX,
-                                self.prefix.title(),
+                                to_pascal_case(self.prefix),
                                 self.connection_field.type.__name__
                             ),
                             params